Split widgets into submodules and sub-submodules
